RAG_Application for text data/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
huggingface-hub
sentence-transformers
transformers
torch
pypdf
